+++ components/data-grid/stories/arg-types.js
@@ -1,0 +1,4 @@
+export const argTypes = {
+	items: { control: { disable: true } },
+	styles: { table: { disable: true } },
+}